move abstract graphics object classes to ImageObjects namespace and separate file
